fix: Correct sidebar layout and highlight style
This commit addresses two user-reported issues:
1.  A bug where the sidebar would disappear on prompt detail pages.
2.  A request to change the sidebar's active link highlight style.

The missing sidebar was caused by the default layout not being
applied to the documents in the `prompts` collection. This has been
fixed by explicitly adding `layout: default` to the front matter of
all prompt files.

The active link highlight style has been updated in
`assets/css/style.scss` to use a grayscale "lift-up" effect
with a box-shadow, as requested by the user.